Move Results component to new file
