+++ src/components/notebooks/dashboard/NotebookDashboard.tsx
@@ -3,7 +3,6 @@ import { DebugModeToggle } from "@/components/debug/DebugModeToggle";
 import { LoadingState } from "@/components/loading/LoadingState";
 import { RuntLogoSmall } from "@/components/logo/RuntLogoSmall";
 import { GitCommitHash } from "@/components/notebook/GitCommitHash";
-import { NotebookGrid } from "@/components/notebooks/NotebookListView";
 import { SimpleUserProfile } from "@/components/notebooks/SimpleUserProfile";
 import { TagActions } from "@/components/notebooks/TagActions";
 import { TagBadge } from "@/components/notebooks/TagBadge";
@@ -16,69 +15,14 @@ import { Input } from "@/components/ui/input";
 import { useNotebooks } from "@/hooks/use-notebooks";
 import { trpcQueryClient } from "@/lib/trpc-client";
 import { useQuery } from "@tanstack/react-query";
-import {
-  Clock,
-  Grid3X3,
-  List,
-  Plus,
-  Search,
-  Tag,
-  User,
-  Users,
-  X,
-} from "lucide-react";
+import { Grid3X3, List, Plus, Search, User, Users } from "lucide-react";
 import React, { useMemo } from "react";
 import { useCreateNotebookAndNavigate, useDashboardParams } from "./helpers";
+import { NoResults, Results } from "./Results";
 
 const DebugNotebooks = React.lazy(() =>
   import("../DebugNotebooks").then((mod) => ({ default: mod.DebugNotebooks }))
 );
-
-function useSmartDefaultFilter({
-  allNotebooks,
-}: {
-  allNotebooks: NotebookProcessed[];
-}) {
-  const { activeFilter, setActiveFilter } = useDashboardParams();
-
-  // Check if user has named notebooks
-  const hasNamedNotebooks = useMemo(() => {
-    return allNotebooks.some(
-      (n) =>
-        n.myPermission === "OWNER" && n.title && !n.title.startsWith("Untitled")
-    );
-  }, [allNotebooks]);
-
-  // Set smart default filter when data first loads
-  const hasNotebooks = allNotebooks.length > 0;
-  React.useEffect(() => {
-    if (hasNotebooks && activeFilter === "named") {
-      const hasScratchNotebooks = allNotebooks.some(
-        (n) =>
-          n.myPermission === "OWNER" &&
-          (!n.title || n.title.startsWith("Untitled"))
-      );
-      const hasSharedNotebooks = allNotebooks.some(
-        (n) => n.myPermission === "WRITER"
-      );
-
-      // Smart default: named > scratch > shared
-      if (hasNamedNotebooks) {
-        // Stay on named
-      } else if (hasScratchNotebooks) {
-        setActiveFilter("scratch");
-      } else if (hasSharedNotebooks) {
-        setActiveFilter("shared");
-      }
-    }
-  }, [
-    hasNotebooks,
-    activeFilter,
-    hasNamedNotebooks,
-    allNotebooks,
-    setActiveFilter,
-  ]);
-}
 
 export const NotebookDashboard: React.FC = () => {
   const debug = useDebug();
@@ -129,7 +73,7 @@ export const NotebookDashboard: React.FC = () => {
           {isLoading && <LoadingState message="Loading notebooks..." />}
           {debug.enabled && <DebugNotebooks notebooks={filteredNotebooks} />}
 
-          {!isLoading && filteredNotebooks.length === 0 && <EmptyState />}
+          {!isLoading && filteredNotebooks.length === 0 && <NoResults />}
 
           {!isLoading && filteredNotebooks.length > 0 && (
             <Results
@@ -310,6 +254,8 @@ function Filters({ allNotebooks }: { allNotebooks: NotebookProcessed[] }) {
   );
 }
 
+// ---
+
 function TopHeader() {
   const { viewMode, setViewMode } = useDashboardParams();
   const createNotebook = useCreateNotebookAndNavigate();
@@ -380,220 +326,48 @@ function SearchInput() {
   );
 }
 
-function EmptyState() {
-  const { searchQuery, activeFilter } = useDashboardParams();
-  const createNotebook = useCreateNotebookAndNavigate();
-  const isSearching = searchQuery.trim();
-
-  return (
-    <div className="py-12 text-center">
-      <div className="mb-4 text-gray-400">
-        <Users className="mx-auto h-16 w-16" />
-      </div>
-      <h3 className="mb-2 text-lg font-medium text-gray-900">
-        {isSearching
-          ? "No notebooks found"
-          : activeFilter === "scratch"
-            ? "No scratch notebooks yet"
-            : activeFilter === "named"
-              ? "No named notebooks yet"
-              : "No shared notebooks yet"}
-      </h3>
-      <p className="mb-6 text-gray-600">
-        {isSearching
-          ? `No notebooks match "${searchQuery}"`
-          : activeFilter === "scratch"
-            ? "Start experimenting - create, iterate, and promote the good ones."
-            : activeFilter === "named"
-              ? "Name your scratch notebooks to promote them here."
-              : "No notebooks have been shared with you yet."}
-      </p>
-      {!isSearching && activeFilter !== "shared" && (
-        <Button onClick={createNotebook}>
-          <Plus className="mr-2 h-4 w-4" />
-          {activeFilter === "scratch"
-            ? "Start Experimenting"
-            : "Create Your First Notebook"}
-        </Button>
-      )}
-    </div>
-  );
-}
-
-function Results({
-  refetch,
-  recentScratchNotebooks,
-  namedNotebooks,
-  filteredNotebooks,
+function useSmartDefaultFilter({
+  allNotebooks,
 }: {
-  refetch: () => void;
-  recentScratchNotebooks: NotebookProcessed[];
-  namedNotebooks: NotebookProcessed[];
-  filteredNotebooks: NotebookProcessed[];
+  allNotebooks: NotebookProcessed[];
 }) {
-  const { searchQuery, activeFilter, viewMode } = useDashboardParams();
-  const isSearching = searchQuery.trim();
+  const { activeFilter, setActiveFilter } = useDashboardParams();
 
-  return (
-    <div className="space-y-8">
-      <ActiveFilters />
+  // Check if user has named notebooks
+  const hasNamedNotebooks = useMemo(() => {
+    return allNotebooks.some(
+      (n) =>
+        n.myPermission === "OWNER" && n.title && !n.title.startsWith("Untitled")
+    );
+  }, [allNotebooks]);
 
-      {/* Search Results Section (prioritized when searching) */}
-      {isSearching && (
-        <section>
-          <div className="mb-4 flex items-center">
-            <Users className="mr-2 h-5 w-5 text-gray-500" />
-            <h2 className="text-lg font-semibold text-gray-900">
-              Search Results
-            </h2>
-            <Badge variant="secondary" className="ml-2">
-              {filteredNotebooks.length}
-            </Badge>
-          </div>
-          <NotebookGrid
-            notebooks={filteredNotebooks}
-            viewMode={viewMode}
-            onUpdate={() => refetch()}
-          />
-        </section>
-      )}
+  // Set smart default filter when data first loads
+  const hasNotebooks = allNotebooks.length > 0;
+  React.useEffect(() => {
+    if (hasNotebooks && activeFilter === "named") {
+      const hasScratchNotebooks = allNotebooks.some(
+        (n) =>
+          n.myPermission === "OWNER" &&
+          (!n.title || n.title.startsWith("Untitled"))
+      );
+      const hasSharedNotebooks = allNotebooks.some(
+        (n) => n.myPermission === "WRITER"
+      );
 
-      {/* Named Notebooks Section (for named filter when not searching) */}
-      {!isSearching &&
-        activeFilter === "named" &&
-        namedNotebooks.length > 0 && (
-          <section>
-            <div className="mb-4 flex items-center">
-              <Tag className="mr-2 h-5 w-5 text-gray-500" />
-              <h2 className="text-lg font-semibold text-gray-900">
-                My Notebooks
-              </h2>
-              <Badge variant="secondary" className="ml-2">
-                {namedNotebooks.length}
-              </Badge>
-            </div>
-            <NotebookGrid
-              notebooks={namedNotebooks}
-              viewMode={viewMode}
-              onUpdate={() => refetch()}
-            />
-          </section>
-        )}
-
-      {/* Recent Scratch Work Section (for scratch filter when not searching) */}
-      {!isSearching &&
-        activeFilter === "scratch" &&
-        recentScratchNotebooks.length > 0 && (
-          <section>
-            <div className="mb-4 flex items-center">
-              <Clock className="mr-2 h-5 w-5 text-gray-500" />
-              <h2 className="text-lg font-semibold text-gray-900">
-                Recent Scratch Work
-              </h2>
-              <Badge variant="secondary" className="ml-2">
-                {recentScratchNotebooks.length}
-              </Badge>
-            </div>
-            <NotebookGrid
-              notebooks={recentScratchNotebooks}
-              viewMode={viewMode}
-              onUpdate={() => refetch()}
-            />
-          </section>
-        )}
-
-      {/* All Results (for shared filter or fallback when not searching) */}
-      {!isSearching &&
-        (activeFilter === "shared" ||
-          (activeFilter === "scratch" &&
-            !recentScratchNotebooks.length &&
-            filteredNotebooks.length > 0) ||
-          (activeFilter === "named" &&
-            !namedNotebooks.length &&
-            filteredNotebooks.length > 0)) && (
-          <section>
-            <div className="mb-4 flex items-center">
-              <Users className="mr-2 h-5 w-5 text-gray-500" />
-              <h2 className="text-lg font-semibold text-gray-900">
-                {activeFilter === "shared"
-                  ? "Shared with Me"
-                  : activeFilter === "scratch"
-                    ? "All Scratch Work"
-                    : "All My Notebooks"}
-              </h2>
-              <Badge variant="secondary" className="ml-2">
-                {filteredNotebooks.length}
-              </Badge>
-            </div>
-            <NotebookGrid
-              notebooks={filteredNotebooks}
-              viewMode={viewMode}
-              onUpdate={() => refetch()}
-            />
-          </section>
-        )}
-    </div>
-  );
-}
-
-function ActiveFilters() {
-  const { searchQuery, setSearchQuery, selectedTagName, setSelectedTag } =
-    useDashboardParams();
-
-  const trpc = useTrpc();
-  const { data: tagsData } = useQuery(trpc.tags.queryOptions());
-  const isSearching = searchQuery.trim();
-
-  return (
-    <>
-      {/* Active Filters Indicator */}
-      {(isSearching || selectedTagName) && (
-        <div className="flex flex-wrap items-center gap-2 rounded-lg border bg-gray-50 p-3">
-          <span className="text-sm text-gray-600">Active filters:</span>
-          {isSearching && (
-            <div className="flex items-center gap-1 rounded-full bg-blue-100 px-2 py-1 text-xs">
-              <Search className="h-3 w-3" />
-              <span>"{searchQuery}"</span>
-              <button
-                onClick={() => setSearchQuery("")}
-                className="text-gray-500 hover:text-gray-700"
-              >
-                <X className="h-3 w-3" />
-              </button>
-            </div>
-          )}
-          {selectedTagName && tagsData && (
-            <div className="flex items-center gap-1">
-              {(() => {
-                const selectedTag = tagsData.find(
-                  (t) => t.name === selectedTagName
-                );
-                return selectedTag ? (
-                  <div className="flex items-center gap-1 rounded-full bg-gray-200 px-2 py-1 text-xs">
-                    <Tag className="h-3 w-3" />
-                    <TagBadge tag={selectedTag} className="text-xs" />
-                    <button
-                      onClick={() => setSelectedTag("")}
-                      className="text-gray-500 hover:text-gray-700"
-                    >
-                      <X className="h-3 w-3" />
-                    </button>
-                  </div>
-                ) : null;
-              })()}
-            </div>
-          )}
-          <button
-            onClick={() => {
-              setSearchQuery("");
-              setSelectedTag("");
-            }}
-            className="ml-auto text-xs text-blue-600 hover:text-blue-800"
-          >
-            Clear all filters
-          </button>
-        </div>
-      )}
-    </>
-  );
+      // Smart default: named > scratch > shared
+      if (hasNamedNotebooks) {
+        // Stay on named
+      } else if (hasScratchNotebooks) {
+        setActiveFilter("scratch");
+      } else if (hasSharedNotebooks) {
+        setActiveFilter("shared");
+      }
+    }
+  }, [
+    hasNotebooks,
+    activeFilter,
+    hasNamedNotebooks,
+    allNotebooks,
+    setActiveFilter,
+  ]);
 }
